services/scanner/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.25
jinja2==3.1.3
